Refactor: extracted common methods between AMFValidatorPlugins in amf-validation and amf-custom-validation
